Resolve part of CR requests
