+++ automation/qbft/runner/runner.go
@@ -2,12 +2,13 @@ package runner
 
 import (
 	"context"
+	"testing"
 
 	"go.uber.org/zap"
 )
 
 // Start starts the runner.
-func Start(logger *zap.Logger, scenario Scenario, bootstrapper Bootstrapper) {
+func Start(t *testing.T, logger *zap.Logger, scenario Scenario, bootstrapper Bootstrapper) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
@@ -16,28 +17,13 @@ func Start(logger *zap.Logger, scenario Scenario, bootstrapper Bootstrapper) {
 		logger.Panic("could not bootstrap scenario", zap.Error(err))
 	}
 
-	if err := run(logger, scenario, sctx); err != nil {
+	scenario.ApplyCtx(sctx)
+
+	logger.Info("all resources were created, starting execution of the scenario")
+
+	if err := scenario.Run(t); err != nil {
 		logger.Panic("could not run scenario", zap.Error(err))
-	}
-}
-
-func run(logger *zap.Logger, scenario Scenario, sctx *ScenarioContext) error {
-	logger.Info("all resources were created, starting pre-execution of the scenario")
-	if err := scenario.PreExecution(sctx); err != nil {
-		return err
-	}
-
-	logger.Info("executing scenario")
-	if err := scenario.Execute(sctx); err != nil {
-		return err
-	}
-
-	logger.Info("running post-execution of the scenario")
-	if err := scenario.PostExecution(sctx); err != nil {
-		return err
 	}
 
 	logger.Info("done")
-
-	return nil
 }
