+++ automation/qbft/tests/all_test.go
@@ -18,6 +18,6 @@ func Test_Automation_QBFTScenarios(t *testing.T) {
 
 	for _, s := range scenariosToRun {
 		scenario := scenarios.NewScenario(s, logger)
-		runner.Start(logger, scenario, scenarios.QBFTScenarioBootstrapper())
+		runner.Start(t, logger, scenario, scenarios.QBFTScenarioBootstrapper())
 	}
 }
